Rename draft.py to display.py, deleted Tetris_Turtle.py
